requirements.txt
ruamel.yaml
pexpect
setuptools
python-ambariclient
satori-rtm-sdk